Feat: Refactor showcases into a single page with menu
