+++ src/react/components/NavBar.tsx
@@ -11,11 +11,10 @@ function NavBar() {
   return (
     <nav>
       <ul>
-        {link("/", "Home")}
-        {link("/items", "Items")}
-        {link("/use-action-state-showcase", "Showcase")}
-        {link("/use-form-status-showcase", "useFormStatus")}
-      </ul>
+  {link("/", "Home")}
+  {link("/items", "Items")}
+  {link("/showcase", "Showcase")}
+</ul>
     </nav>
   );
 }
