+++ src/react/routes.tsx
@@ -9,8 +9,8 @@ import ItemCreate from "./pages/ItemCreate";
 import ItemEdit from "./pages/ItemEdit";
 import ItemList from "./pages/ItemList";
 import ItemShow from "./pages/ItemShow";
-import UseActionStateShowcase from './pages/UseActionStateShowcase';
-import UseFormStatusShowcase from './pages/UseFormStatusShowcase';
+
+import ShowcasePage from './pages/Showcase';
 
 import "./index.css";
 
@@ -28,13 +28,10 @@ export default [
         element: <Home />,
       },
       {
-        path: "/use-action-state-showcase",
-        element: <UseActionStateShowcase />,
+       path: "/showcase",
+       element: <ShowcasePage />,
       },
-      {
-        path: "/use-form-status-showcase",
-        element: <UseFormStatusShowcase />,
-    },
+      
       {
         path: "/items",
         element: (
